Simplify greeting API
