+++ docs/dxp/7.x/en/liferay-internals/fundamentals/apis-as-osgi-services/liferay-p9g2.zip/p9g2-impl/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Acme P9G2 Greeting
-Bundle-SymbolicName: com.liferay.learn.p9g2.impl
+Bundle-Name: Acme P9G2 Greeting Implementation
+Bundle-SymbolicName: com.acme.p9g2.impl
 Bundle-Version: 1.0.0
