+++ docs/dxp/7.x/en/liferay-internals/fundamentals/module-projects/liferay-k8s2.zip/k8s2-api/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Acme K8S2 Greeting API
 Bundle-SymbolicName: com.acme.k8s2.api
 Bundle-Version: 1.0.0
-Export-Package: com.acme.k8s2.greeting
+Export-Package: com.acme.k8s2
