+++ docs/dxp/7.x/en/liferay-internals/fundamentals/module-projects/liferay-k8s2.zip/k8s2-api/src/main/java/com/acme/k8s2/Greeting.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.acme.j1h1.greeting;
+package com.acme.k8s2;
 
 import aQute.bnd.annotation.ProviderType;
