+++ docs/dxp/7.x/en/liferay-internals/fundamentals/using-an-osgi-service/liferay-j1h1.zip/j1h1-client/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Acme J1H1 Greeting Command
-Bundle-SymbolicName: com.liferay.learn.j1h1.client
+Bundle-Name: Acme J1H1 Greeting Client
+Bundle-SymbolicName: com.acme.j1h1.client
 Bundle-Version: 1.0.0
