+++ docs/dxp/7.x/en/liferay-internals/fundamentals/using-an-osgi-service/liferay-j1h1.zip/j1h1-impl/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Acme J1H1 Greeting
-Bundle-SymbolicName: com.liferay.learn.j1h1.impl
+Bundle-Name: Acme J1H1 Greeting Implementation
+Bundle-SymbolicName: com.acme.j1h1.impl
 Bundle-Version: 1.0.0
